requirements.txt
openai-whisper
torch>=2.0.0
numpy>=1.24.0
spacy>=3.0.0
